Code cleanup

+++ include/cellular-evolution/utilities/evo_utils.h
@@ -20,9 +20,9 @@ void LABtoRGB(int &r_r, int &g_r, int &b_r, int const l, int const a, int const 
     g_t = x * -0.9689f + y * 1.8758f + z * 0.0415f;
     b_t = x * 0.0557f + y * -0.2040f + z * 1.0570f;
 
-    r_t = (float) ((r_t > 0.0031308f) ? (1.055f * std::pow(r_t, 1.0f / 2.4f) - 0.055f) : 12.92f * r_t);
-    g_t = (float) ((g_t > 0.0031308f) ? (1.055f * std::pow(g_t, 1.0f / 2.4f) - 0.055f) : 12.92f * g_t);
-    b_t = (float) ((b_t > 0.0031308f) ? (1.055f * std::pow(b_t, 1.0f / 2.4f) - 0.055f) : 12.92f * b_t);
+    r_t = (r_t > 0.0031308f) ? (1.055f * std::pow(r_t, 1.0f / 2.4f) - 0.055f) : 12.92f * r_t;
+    g_t = (g_t > 0.0031308f) ? (1.055f * std::pow(g_t, 1.0f / 2.4f) - 0.055f) : 12.92f * g_t;
+    b_t = (b_t > 0.0031308f) ? (1.055f * std::pow(b_t, 1.0f / 2.4f) - 0.055f) : 12.92f * b_t;
 
     r_r = (int) (std::fmax(0, std::fmin(1, r_t)) * 255);
     g_r = (int) (std::fmax(0, std::fmin(1, g_t)) * 255);
